add test to make file

+++ memtable/artrep.cc
@@ -41,7 +41,7 @@ public:
 
   // Returns true iff an entry that compares equal to key is in the list.
  bool Contains(const char* key) const override {
-   return skip_list_.Get(key, strlen(key)) != nullptr;
+   return skip_list_.Get(key, static_cast<int>(strlen(key))) != nullptr;
  }
 
  size_t ApproximateMemoryUsage() override {
@@ -51,7 +51,8 @@ public:
 
  void Get(const LookupKey& k, void* callback_args,
           bool (*callback_func)(void* arg, const char* entry)) override {
-   const char* value = skip_list_.Get(k.user_key().data(), k.user_key().size());
+   const char* value = skip_list_.Get(k.user_key().data(),
+                                      static_cast<int>(k.user_key().size()));
    if (value != nullptr) {
      callback_func(callback_args, value);
    }
@@ -93,9 +94,9 @@ public:
       if (memtable_key != nullptr) {
         uint32_t l = 0;
         const char* k = GetVarint32Ptr(memtable_key, memtable_key + 5, &l);
-        iter_.Seek(k, l - 8);
+        iter_.Seek(k, static_cast<int>(l) - 8);
       } else {
-        iter_.Seek(user_key.data(), user_key.size() - 8);
+        iter_.Seek(user_key.data(), static_cast<int>(user_key.size()) - 8);
       }
     }
 
@@ -106,7 +107,8 @@ public:
         const char* k = GetVarint32Ptr(memtable_key, memtable_key + 5, &l);
         iter_.SeekForPrev(k, l - 8);
       } else {
-        iter_.SeekForPrev(user_key.data(), user_key.size() - 8);
+        iter_.SeekForPrev(user_key.data(),
+                          static_cast<int>(user_key.size()) - 8);
       }
     }
